Refactor codeMapMouseEvent.ts, Highlight labels when hovering building #1529
